Removes the loadCSS preload polyfill
https://github.com/filamentgroup/loadCSS#changes-in-version-30-no-more-preload-polyfill

> As of version 3.0, we no longer support or include a polyfill for a
rel=preload markup pattern. This is because we have since determined
that the markup pattern described at the top of this readme is
simpler and better for performance, while the former preload pattern
could sometimes conflict with resource priorities in ways that aren't
helpful for loading CSS in a non-blocking way.