Changes the method with which we hide the sprite
I recently hit an issue with an SVG file that wasn't rendering correctly because it contains a `url()` function pointing to a linearGradient element. I found that hiding the sprite sheet with `display:none` was the culprit. I've changed the way we hide the sprite sheet to match the way WP core handles them.